test_time_scaling/pyproject.toml
[project]
name = "test-time-scaling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "accelerate==1.0.1",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.10.11",
    "aiosignal==1.3.1",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "datasets==3.1.0",
    "dill==0.3.8",
    "filelock==3.16.1",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "hf-xet==1.2.0",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "markupsafe==2.1.5",
    "matplotlib>=3.7.5",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "networkx==3.1",
    "numpy==1.24.4",
    "nvidia-cublas-cu12==12.1.3.1",
    "nvidia-cuda-cupti-cu12==12.1.105",
    "nvidia-cuda-nvrtc-cu12==12.1.105",
    "nvidia-cuda-runtime-cu12==12.1.105",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.0.2.54",
    "nvidia-curand-cu12==10.3.2.106",
    "nvidia-cusolver-cu12==11.4.5.107",
    "nvidia-cusparse-cu12==12.1.0.106",
    "nvidia-nccl-cu12==2.20.5",
    "nvidia-nvjitlink-cu12==12.9.86",
    "nvidia-nvtx-cu12==12.1.105",
    "openai>=2.2.0",
    "packaging==25.0",
    "pandas==2.0.3",
    "propcache==0.2.0",
    "psutil==7.1.2",
    "pyarrow==17.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2024.11.6",
    "requests==2.32.4",
    "safetensors==0.5.3",
    "six==1.17.0",
    "sympy==1.13.3",
    "tokenizers==0.20.3",
    "torch==2.4.1",
    "tqdm==4.67.1",
    "transformers==4.46.3",
    "triton==3.0.0",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.2.3",
    "xxhash==3.6.0",
    "yarl==1.15.2",
]
